Add comments and skipNextIteration support
This is necessary in order to synchronize the iterations
Also identified a bug, where in the token implementation, in readyToProgress it would trigger a next time step for every incoming message, i.e. multiple times. Workaround, but not so nice.
Open problem: The token circulation should start again with first network at the beginning of each time step. Currently not the case.